Allow insertImage without requiring uploadImage
