+++ packages/ckeditor5-image/tests/manual/imageinsertviaurl.md
@@ -1,23 +1,17 @@
-## Image insert via URL
+## Image insert via URL when image upload is not available
 
-1. Click on the arrow button in `insertImage` plugin to reveal the image insert panel.
+1. Click on the plugin icon to reveal the image insert panel.
 1. Paste the URL to the input (eg: `https://ckeditor.com/docs/ckeditor5/latest/assets/img/malta.jpg`).
 1. Click `Insert` button.
 
-## Image replace via URL
+## Image replace via URL when image upload is not available
 
 1. Click on the image in the editor.
-1. Click on the arrow button in `insertImage` plugin to reveal the image insert panel.
+1. Click on the plugin icon to reveal the image insert panel.
 1. Edit the value of the input.
 1. Click `Update` button.
 
-## Image insert via integrations
-
-1. In the **Editor 2** click on the arrow button in `insertImage` plugin to reveal the image insert panel.
-1. Click on the **CKFinder** button.
-1. Choose image and confirm.
-
-## Image insert via pasting URL into editor
+## Image insert via pasting URL into editor when image upload is not available
 
 1. Copy image URL to clipboard (eg: `https://ckeditor.com/docs/ckeditor5/latest/assets/img/malta.jpg`).
 1. Paste the URL to the editor.
