+++ packages/ckeditor5-image/tests/manual/imageinsertviaurlwithupload.html
@@ -11,7 +11,7 @@
 </style>
 
 <div id="editor">
-	<h2>Image upload via URL without image upload option</h2>
+	<h2>Image upload via URL with image upload and CKFinder integration</h2>
 	<p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla finibus consequat placerat. Vestibulum id tellus et mauris sagittis tincidunt quis id mauris. Curabitur consectetur lectus sit amet tellus mattis, non lobortis leo interdum.</p>
 </div>
