+++ packages/ckeditor5-image/tests/manual/imageinsertviaurlwithupload.md
@@ -1,0 +1,23 @@
+## Image insert via URL when image upload is available
+
+1. Click on the arrow button in `insertImage` plugin to reveal the image insert panel.
+1. Paste the URL to the input (eg: `https://ckeditor.com/docs/ckeditor5/latest/assets/img/malta.jpg`).
+1. Click `Insert` button.
+
+## Image replace via URL when image upload is available
+
+1. Click on the image in the editor.
+1. Click on the arrow button in `insertImage` plugin to reveal the image insert panel.
+1. Edit the value of the input.
+1. Click `Update` button.
+
+## Image insert via integrations when image upload is available
+
+1. In the **Editor 2** click on the arrow button in `insertImage` plugin to reveal the image insert panel.
+1. Click on the **CKFinder** button.
+1. Choose image and confirm.
+
+## Image insert via pasting URL into editor when image upload is available
+
+1. Copy image URL to clipboard (eg: `https://ckeditor.com/docs/ckeditor5/latest/assets/img/malta.jpg`).
+1. Paste the URL to the editor.
